+++ modules/perception/lidar/tools/exporter/msg_exporter.cc
@@ -24,44 +24,41 @@
 #include "absl/strings/str_split.h"
 #include "pcl/io/pcd_io.h"
 
+#include "modules/transform/proto/transform.pb.h"
+
 #include "cyber/common/file.h"
 #include "cyber/common/log.h"
-#include "modules/transform/proto/transform.pb.h"
 
 namespace apollo {
 namespace perception {
 namespace lidar {
 
 MsgExporter::MsgExporter(std::shared_ptr<apollo::cyber::Node> node,
-                         const std::vector<std::string> channels,
-                         const std::vector<std::string> child_frame_ids) {
-  _cyber_node = node;
-  _channels = channels;
-  _child_frame_ids = child_frame_ids;
+                         const std::vector<std::string>       channels,
+                         const std::vector<std::string>       child_frame_ids) {
+  _cyber_node          = node;
+  _channels            = channels;
+  _child_frame_ids     = child_frame_ids;
   _localization_method = "perception_localization_100hz";
-  auto create_folder = [](const std::string& folder) {
-    if (cyber::common::DirectoryExists(folder)) {
-      cyber::common::DeleteFile(folder);
-    }
+  auto create_folder   = [](const std::string& folder) {
+    if (cyber::common::DirectoryExists(folder)) { cyber::common::DeleteFile(folder); }
     cyber::common::CreateDir(folder);
   };
   // bind channels
   for (std::size_t i = 0; i < _channels.size(); ++i) {
-    const auto& channel = _channels[i];
+    const auto& channel        = _channels[i];
     const auto& child_frame_id = _child_frame_ids[i];
-    std::string folder = TransformChannelToFolder(channel);
+    std::string folder         = TransformChannelToFolder(channel);
     if (IsLidar(channel)) {
-      _cyber_node->CreateReader<PcMsg>(
-          channel,
-          std::bind(&MsgExporter::PointCloudMessageHandler, this,
-                    std::placeholders::_1, channel, child_frame_id, folder));
+      _cyber_node->CreateReader<PcMsg>(channel, std::bind(&MsgExporter::PointCloudMessageHandler,
+                                                          this, std::placeholders::_1, channel,
+                                                          child_frame_id, folder));
       create_folder(folder);
       std::cout << "Bind lidar channel: " << channel << std::endl;
     } else if (IsCamera(channel)) {
-      _cyber_node->CreateReader<ImgMsg>(
-          channel,
-          std::bind(&MsgExporter::ImageMessageHandler, this,
-                    std::placeholders::_1, channel, child_frame_id, folder));
+      _cyber_node->CreateReader<ImgMsg>(channel, std::bind(&MsgExporter::ImageMessageHandler, this,
+                                                           std::placeholders::_1, channel,
+                                                           child_frame_id, folder));
       create_folder(folder);
       std::cout << "Bind camera channel: " << channel << std::endl;
     } else {
@@ -70,33 +67,33 @@ MsgExporter::MsgExporter(std::shared_ptr<apollo::cyber::Node> node,
   }
 }
 
-void MsgExporter::ImageMessageHandler(
-    const std::shared_ptr<const ImgMsg>& img_msg, const std::string& channel,
-    const std::string& child_frame_id, const std::string& folder) {
+void MsgExporter::ImageMessageHandler(const std::shared_ptr<const ImgMsg>& img_msg,
+                                      const std::string&                   channel,
+                                      const std::string&                   child_frame_id,
+                                      const std::string&                   folder) {
   double timestamp = img_msg->measurement_time();
   // std::cout << "Receive image message from channel: " << channel <<
   //    " at time: " << FORMAT_TIMESTAMP(timestamp) << std::endl;
-  const unsigned char* data =
-      reinterpret_cast<const unsigned char*>(img_msg->data().data());
+  const unsigned char* data       = reinterpret_cast<const unsigned char*>(img_msg->data().data());
   const unsigned char* range_data = nullptr;
   // const unsigned char* range_data = is_stereo_camera(channel) ?
   //    reinterpret_cast<const unsigned char*>(img_msg->disparity_data().data())
   //    : nullptr;
-  std::size_t width = img_msg->width();
-  std::size_t height = img_msg->height();
+  std::size_t     width  = img_msg->width();
+  std::size_t     height = img_msg->height();
   Eigen::Matrix4d pose;
   if (QuerySensorToWorldPose(timestamp, child_frame_id, &pose)) {
     SaveImage(data, range_data, width, height, timestamp, folder);
     SavePose(pose, timestamp, folder);
   } else {
-    std::cout << "Failed to query camera pose, child_frame_id "
-              << child_frame_id << std::endl;
+    std::cout << "Failed to query camera pose, child_frame_id " << child_frame_id << std::endl;
   }
 }
 
-void MsgExporter::PointCloudMessageHandler(
-    const std::shared_ptr<const PcMsg>& cloud_msg, const std::string& channel,
-    const std::string& child_frame_id, const std::string& folder) {
+void MsgExporter::PointCloudMessageHandler(const std::shared_ptr<const PcMsg>& cloud_msg,
+                                           const std::string&                  channel,
+                                           const std::string&                  child_frame_id,
+                                           const std::string&                  folder) {
   double timestamp = cloud_msg->measurement_time();
   // std::cout << "Receive point cloud message from channel: " << channel <<
   //    " at time: " << FORMAT_TIMESTAMP(timestamp) << std::endl;
@@ -105,16 +102,13 @@ void MsgExporter::PointCloudMessageHandler(
     cloud.resize(cloud_msg->point_size());
     for (int i = 0; i < cloud_msg->point_size(); ++i) {
       const apollo::drivers::PointXYZIT& pt = cloud_msg->point(i);
-      if (std::isnan(pt.x()) || std::isnan(pt.y()) || std::isnan(pt.z())) {
+      if (std::isnan(pt.x()) || std::isnan(pt.y()) || std::isnan(pt.z())) { continue; }
+      if (std::fabs(pt.x()) > 1e3 || std::fabs(pt.y()) > 1e3 || std::fabs(pt.z()) > 1e3) {
         continue;
       }
-      if (std::fabs(pt.x()) > 1e3 || std::fabs(pt.y()) > 1e3 ||
-          std::fabs(pt.z()) > 1e3) {
-        continue;
-      }
-      cloud[i].x = pt.x();
-      cloud[i].y = pt.y();
-      cloud[i].z = pt.z();
+      cloud[i].x         = pt.x();
+      cloud[i].y         = pt.y();
+      cloud[i].z         = pt.z();
       cloud[i].intensity = static_cast<unsigned char>(pt.intensity());
       cloud[i].timestamp = static_cast<double>(pt.timestamp()) * 1e-9;
     }
@@ -124,14 +118,13 @@ void MsgExporter::PointCloudMessageHandler(
     SavePointCloud(cloud, timestamp, folder);
     SavePose(pose, timestamp, folder);
   } else {
-    std::cout << "Failed to query lidar pose, child_frame_id " << child_frame_id
-              << std::endl;
+    std::cout << "Failed to query lidar pose, child_frame_id " << child_frame_id << std::endl;
   }
 }
 
-bool MsgExporter::SavePointCloud(
-    const pcl::PointCloud<PCLPointXYZIT>& point_cloud, double timestamp,
-    const std::string& folder) {
+bool MsgExporter::SavePointCloud(const pcl::PointCloud<PCLPointXYZIT>& point_cloud,
+                                 double                                timestamp,
+                                 const std::string&                    folder) {
   static char path[500];
   snprintf(path, sizeof(path), "%s/%.6f.pcd", folder.c_str(), timestamp);
   pcl::PCDWriter writer;
@@ -140,10 +133,12 @@ bool MsgExporter::SavePointCloud(
 }
 
 bool MsgExporter::SaveImage(const unsigned char* color_image,
-                            const unsigned char* range_image, std::size_t width,
-                            std::size_t height, double timestamp,
-                            const std::string& folder) {
-  cv::Mat image(static_cast<int>(height), static_cast<int>(width), CV_8UC1,
+                            const unsigned char* range_image,
+                            std::size_t          width,
+                            std::size_t          height,
+                            double               timestamp,
+                            const std::string&   folder) {
+  cv::Mat     image(static_cast<int>(height), static_cast<int>(width), CV_8UC1,
                 const_cast<unsigned char*>(color_image));
   static char path[500];
   snprintf(path, sizeof(path), "%s/%.6f.jpg", folder.c_str(), timestamp);
@@ -164,33 +159,29 @@ bool MsgExporter::SaveImage(const unsigned char* color_image,
   return true;
 }
 
-bool MsgExporter::QuerySensorToWorldPose(double timestamp,
+bool MsgExporter::QuerySensorToWorldPose(double             timestamp,
                                          const std::string& child_frame_id,
-                                         Eigen::Matrix4d* pose) {
+                                         Eigen::Matrix4d*   pose) {
   Eigen::Matrix4d sensor2novatel_extrinsics = Eigen::Matrix4d::Identity();
   Eigen::Matrix4d novatel2world_pose;
   // query sensor to novatel extrinsics
   if (child_frame_id != "null" &&
-      !QueryPose(timestamp, "novatel", child_frame_id,
-                 &sensor2novatel_extrinsics)) {
+      !QueryPose(timestamp, "novatel", child_frame_id, &sensor2novatel_extrinsics)) {
     return false;
   }
   // query novatel to world pose
-  if (!QueryPose(timestamp, "world", _localization_method,
-                 &novatel2world_pose)) {
-    return false;
-  }
+  if (!QueryPose(timestamp, "world", _localization_method, &novatel2world_pose)) { return false; }
   *pose = novatel2world_pose * sensor2novatel_extrinsics;
   return true;
 }
 
-bool MsgExporter::QueryPose(double timestamp, const std::string& frame_id,
+bool MsgExporter::QueryPose(double             timestamp,
+                            const std::string& frame_id,
                             const std::string& child_frame_id,
-                            Eigen::Matrix4d* pose) {
+                            Eigen::Matrix4d*   pose) {
   cyber::Time query_time(timestamp);
   std::string err_string;
-  if (!tf2_buffer_->canTransform(frame_id, child_frame_id, query_time, 0.05f,
-                                 &err_string)) {
+  if (!tf2_buffer_->canTransform(frame_id, child_frame_id, query_time, 0.05f, &err_string)) {
     AERROR << "Can not find transform. "  //<< FORMAT_TIMESTAMP(timestamp)
            << " frame_id: " << frame_id << " child_frame_id: " << child_frame_id
            << " Error info: " << err_string;
@@ -198,13 +189,11 @@ bool MsgExporter::QueryPose(double timestamp, const std::string& frame_id,
   }
   apollo::transform::TransformStamped stamped_transform;
   try {
-    stamped_transform =
-        tf2_buffer_->lookupTransform(frame_id, child_frame_id, query_time);
-    Eigen::Translation3d translation(
-        stamped_transform.transform().translation().x(),
-        stamped_transform.transform().translation().y(),
-        stamped_transform.transform().translation().z());
-    Eigen::Quaterniond rotation(stamped_transform.transform().rotation().qw(),
+    stamped_transform = tf2_buffer_->lookupTransform(frame_id, child_frame_id, query_time);
+    Eigen::Translation3d translation(stamped_transform.transform().translation().x(),
+                                     stamped_transform.transform().translation().y(),
+                                     stamped_transform.transform().translation().z());
+    Eigen::Quaterniond   rotation(stamped_transform.transform().rotation().qw(),
                                 stamped_transform.transform().rotation().qx(),
                                 stamped_transform.transform().rotation().qy(),
                                 stamped_transform.transform().rotation().qz());
@@ -216,17 +205,18 @@ bool MsgExporter::QueryPose(double timestamp, const std::string& frame_id,
   return true;
 }
 
-bool MsgExporter::SavePose(const Eigen::Matrix4d& pose, double timestamp,
-                           const std::string& folder) {
-  Eigen::Affine3d affine(pose);
+bool MsgExporter::SavePose(const Eigen::Matrix4d& pose,
+                           double                 timestamp,
+                           const std::string&     folder) {
+  Eigen::Affine3d    affine(pose);
   Eigen::Quaterniond quat = (Eigen::Quaterniond)affine.linear();
-  static char path[500];
+  static char        path[500];
   snprintf(path, sizeof(path), "%s/%.6f.pose", folder.c_str(), timestamp);
   std::ofstream fout(path);
   if (fout.is_open()) {
     fout << 0 << " " << std::setprecision(14) << timestamp << " "
-         << affine.translation().transpose() << " " << quat.x() << " "
-         << quat.y() << " " << quat.z() << " " << quat.w() << std::endl;
+         << affine.translation().transpose() << " " << quat.x() << " " << quat.y() << " "
+         << quat.z() << " " << quat.w() << std::endl;
     fout.close();
   }
   return true;
@@ -248,8 +238,8 @@ bool MsgExporter::IsLidar(const std::string& channel) {
 }
 
 std::string MsgExporter::TransformChannelToFolder(const std::string& channel) {
-  const std::vector<std::string> strs = absl::StrSplit(channel, '/');
-  std::string target = "";
+  const std::vector<std::string> strs   = absl::StrSplit(channel, '/');
+  std::string                    target = "";
   for (auto& str : strs) {
     target += str + "_";
   }
